Update mysqltuner.spec for improved scripting and release
The mysqltuner.spec has been updated to fix a typo and enhance version release. Additionally, a better scripting function has been added to automatically insert a shebang if the first line of mysqltuner.pl doesn't start with '#!'. This aims to ensure run-time interpretation of the script using the perl interpreter defined in {_bindir}.